Completed Basic crud operations and basic blog archetecture

+++ public/app.js
@@ -7,6 +7,7 @@
       $scope.createPost = createPost;
       $scope.deletePost = deletePost;
       $scope.editPost = editPost;
+      $scope.updatePost = updatePost;
       // $scope.updatePost = updatePost;
 
         function init(){
@@ -23,7 +24,7 @@
               $scope.posts = posts.data;
             },
               function(err){
-                $scope.msg = "failure"
+              swal("awwwwwww!", "Something failed", "error");
               }
           )}
 
@@ -31,22 +32,40 @@
             console.log(post);
             $http.post("/api/blogPost", post)  .then(
                 function(posts){
+                swal("WO HO HO!", "Created Sucessfully", "success");
               getAllPosts();
               },
                 function(err){
-                  $scope.msg = "creation failed"
+              swal("awwwwwww!", "Something failed", "error");
                 })
         }
 
         function deletePost(id){
-          $http.delete("/api/blogPost/"+id)
-          .then(
-            function(posts){
-          getAllPosts();
-          },
-            function(err){
-              $scope.msg = "deletion failed"
-            })
+          swal({
+            title: "Are you sure?",
+            text: "Once deleted, you will not be able to recover this Blog!",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
+          })
+          .then((willDelete) => {
+            if (willDelete) {
+              $http.delete("/api/blogPost/"+id)
+              .then(
+                function(posts){
+              getAllPosts();
+                swal("WO HO HO!!", "Deleted Sucessfully", "info");
+              },
+                function(err){
+                swal("awwwwwww!", "Something failed", "error");
+              });
+            } else {
+              swal("Your Blog is safe  Whew!");
+                swal("WOAAAAAAH!!", "Your Blog is safe  Whew!", "info");
+            }
+          });
+
+
         }
 
         function editPost(id){
@@ -56,7 +75,19 @@
               $scope.post = post.data;
           },
             function(err){
-              $scope.msg = "couldnt retrieve data"
+              swal("awwwwwww!", "Something failed", "error");
+            })
+        }
+
+        function updatePost(post){
+          $http.put("/api/blogPost/"+post._id,post)
+          .then(
+            function(post){
+                getAllPosts();
+            swal("WO HO HO!!", "updated Sucessfully", "success");
+          },
+            function(err){
+                swal("awwwwwww!", "Something failed", "error");
             })
         }
